+++ src/main/java/com/mobble/mobbleserver/refactor/article/dto/request/ArticleRequestDto.java
@@ -2,7 +2,7 @@ package com.mobble.mobbleserver.refactor.article.dto.request;
 
 import com.mobble.mobbleserver.refactor.article.entity.Article;
 import com.mobble.mobbleserver.refactor.article.entity.ArticleType;
-import com.mobble.mobbleserver.refactor.club.core.entity.Club;
+import com.mobble.mobbleserver.domain.club.core.Club;
 import com.mobble.mobbleserver.domain.member.Member;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
