+++ src/main/java/com/mobble/mobbleserver/refactor/chat/clubChatRoom/dto/response/ClubChatRoomPreviewResponseDto.java
@@ -2,7 +2,7 @@ package com.mobble.mobbleserver.refactor.chat.clubChatRoom.dto.response;
 
 import com.mobble.mobbleserver.refactor.chat.chatMessage.entity.ChatMessage;
 import com.mobble.mobbleserver.refactor.chat.chatRoom.entity.ChatRoom;
-import com.mobble.mobbleserver.refactor.club.core.entity.Club;
+import com.mobble.mobbleserver.domain.club.core.Club;
 
 import java.time.LocalDateTime;
