+++ src/main/java/com/mobble/mobbleserver/refactor/club/clubGround/entity/ClubGround.java
@@ -1,6 +1,6 @@
 package com.mobble.mobbleserver.refactor.club.clubGround.entity;
 
-import com.mobble.mobbleserver.refactor.club.core.entity.Club;
+import com.mobble.mobbleserver.domain.club.core.Club;
 import com.mobble.mobbleserver.refactor.ground.entity.Ground;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
