+++ src/main/java/com/mobble/mobbleserver/refactor/clubImage/entity/ClubImage.java
@@ -1,7 +1,7 @@
 package com.mobble.mobbleserver.refactor.clubImage.entity;
 
 import com.mobble.mobbleserver.common.baseEntity.CreatedAtEntity;
-import com.mobble.mobbleserver.refactor.club.core.entity.Club;
+import com.mobble.mobbleserver.domain.club.core.Club;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
