+++ src/main/java/com/mobble/mobbleserver/refactor/clubMember/dto/request/CreateClubMemberDto.java
@@ -1,6 +1,6 @@
 package com.mobble.mobbleserver.refactor.clubMember.dto.request;
 
-import com.mobble.mobbleserver.refactor.club.core.entity.Club;
+import com.mobble.mobbleserver.domain.club.core.Club;
 import com.mobble.mobbleserver.refactor.clubMember.entity.ClubMember;
 import com.mobble.mobbleserver.refactor.clubMember.entity.ClubMemberRole;
 import com.mobble.mobbleserver.refactor.clubMember.entity.JoinStatus;
