+++ src/test/java/com/mobble/mobbleserver/refactor/clubMember/entity/ClubMemberTest.java
@@ -1,7 +1,7 @@
 package com.mobble.mobbleserver.refactor.clubMember.entity;
 
 import com.mobble.mobbleserver.refactor.article.entity.ArticleType;
-import com.mobble.mobbleserver.refactor.club.core.entity.Club;
+import com.mobble.mobbleserver.domain.club.core.Club;
 import com.mobble.mobbleserver.refactor.clubCategory.entity.ClubCategory;
 import com.mobble.mobbleserver.domain.member.Member;
 import com.mobble.mobbleserver.support.fixture.club.ClubTestFixture;
