+++ src/test/java/com/mobble/mobbleserver/refactor/clubMember/repository/ClubMemberRepositoryImplTest.java
@@ -3,7 +3,7 @@ package com.mobble.mobbleserver.refactor.clubMember.repository;
 import com.mobble.mobbleserver.config.QueryDslConfig;
 import com.mobble.mobbleserver.refactor.chat.chatRoom.entity.ChatRoom;
 import com.mobble.mobbleserver.refactor.chat.clubChatRoom.entity.ClubChatRoom;
-import com.mobble.mobbleserver.refactor.club.core.entity.Club;
+import com.mobble.mobbleserver.domain.club.core.Club;
 import com.mobble.mobbleserver.refactor.clubCategory.entity.ClubCategory;
 import com.mobble.mobbleserver.refactor.clubMember.entity.ClubMember;
 import com.mobble.mobbleserver.refactor.clubMember.entity.ClubMemberRole;
