+++ src/test/java/com/mobble/mobbleserver/refactor/like/articleLike/repository/ArticleLikeRepositoryTest.java
@@ -2,7 +2,7 @@ package com.mobble.mobbleserver.refactor.like.articleLike.repository;
 
 import com.mobble.mobbleserver.config.QueryDslConfig;
 import com.mobble.mobbleserver.refactor.article.entity.Article;
-import com.mobble.mobbleserver.refactor.club.core.entity.Club;
+import com.mobble.mobbleserver.domain.club.core.Club;
 import com.mobble.mobbleserver.refactor.clubCategory.entity.ClubCategory;
 import com.mobble.mobbleserver.refactor.like.articleLike.entity.ArticleLike;
 import com.mobble.mobbleserver.domain.member.Member;
