+++ src/test/java/com/mobble/mobbleserver/refactor/meeting/repository/MeetingRepositoryTest.java
@@ -2,7 +2,7 @@ package com.mobble.mobbleserver.refactor.meeting.repository;
 
 import com.mobble.mobbleserver.config.QueryDslConfig;
 import com.mobble.mobbleserver.infrastructure.persistence.meeting.JpaMeetingRepository;
-import com.mobble.mobbleserver.refactor.club.core.entity.Club;
+import com.mobble.mobbleserver.domain.club.core.Club;
 import com.mobble.mobbleserver.refactor.clubCategory.entity.ClubCategory;
 import com.mobble.mobbleserver.refactor.clubMember.entity.ClubMember;
 import com.mobble.mobbleserver.refactor.clubMember.entity.ClubMemberRole;
